reduce the go type of number columns to just one: uint64 for every no-bigint numeric field (ensuring that the value fills into a int64, data type used by the database). (#84)

+++ api/types.go
@@ -4,13 +4,13 @@ type CreateTokenRequest struct {
 	ID      string `json:"id"`
 	Type    string `json:"type"`
 	Tag     string `json:"tag"`
-	ChainID int64  `json:"chainID"`
+	ChainID uint64 `json:"chainID"`
 }
 
 type GetTokenStatusResponse struct {
 	AtBlock  uint64 `json:"atBlock"`
 	Synced   bool   `json:"synced"`
-	Progress uint64 `json:"progress"`
+	Progress int    `json:"progress"`
 }
 
 type GetTokenResponse struct {
@@ -22,20 +22,20 @@ type GetTokenResponse struct {
 	TotalSupply     string                  `json:"totalSupply"`
 	Name            string                  `json:"name"`
 	Status          *GetTokenStatusResponse `json:"status"`
-	Size            uint32                  `json:"size"`
+	Size            uint64                  `json:"size"`
 	DefaultStrategy uint64                  `json:"defaultStrategy,omitempty"`
 	Tag             string                  `json:"tag,omitempty"`
-	ChainID         int64                   `json:"chainID"`
+	ChainID         uint64                  `json:"chainID"`
 }
 
 type GetTokensItem struct {
 	ID         string `json:"id"`
 	Type       string `json:"type"`
-	StartBlock uint64 `json:"startBlock"`
+	StartBlock int64  `json:"startBlock"`
 	Name       string `json:"name"`
 	Symbol     string `json:"symbol"`
 	Tag        string `json:"tag,omitempty"`
-	ChainID    int    `json:"chainID"`
+	ChainID    uint64 `json:"chainID"`
 }
 
 type GetTokensResponse struct {
@@ -51,8 +51,8 @@ type TokenHoldersResponse struct {
 }
 
 type CreateCensusRequest struct {
-	StrategyID  uint32 `json:"strategyId"`
-	BlockNumber uint32 `json:"blockNumber"`
+	StrategyID  uint64 `json:"strategyId"`
+	BlockNumber uint64 `json:"blockNumber"`
 	Anonymous   bool   `json:"anonymous"`
 }
 
@@ -61,21 +61,21 @@ type CreateCensusResponse struct {
 }
 
 type GetCensusResponse struct {
-	CensusID   uint32 `json:"censusId"`
-	StrategyID uint32 `json:"strategyId"`
+	CensusID   uint64 `json:"censusId"`
+	StrategyID uint64 `json:"strategyId"`
 	MerkleRoot string `json:"merkleRoot"`
 	URI        string `json:"uri"`
-	Size       int32  `json:"size"`
+	Size       uint64 `json:"size"`
 	Weight     string `json:"weight"`
 	Anonymous  bool   `json:"anonymous"`
 }
 
 type GetCensusesResponse struct {
-	Censuses []uint32 `json:"censuses"`
+	Censuses []uint64 `json:"censuses"`
 }
 
 type GetStrategiesResponse struct {
-	Strategies []uint32 `json:"strategies"`
+	Strategies []uint64 `json:"strategies"`
 }
 
 type GetStrategyToken struct {
